Hopefully greatly optimize Color code

+++ src/hooks/SceneTransition/SceneTransitionHelper.cpp
@@ -58,7 +58,7 @@ bool SceneTransitionHelper::BasicPatch(GlobalNamespace::IDifficultyBeatmap* cust
 
     auto beatmapEvents = customBeatmapData->get_beatmapData()->beatmapEventsData->items;
     auto length = beatmapEvents->Length();
-    getLogger().debug("Checking at most %d for ChromaLite notes", length);
+    getLogger().debug("Checking at most %d for ChromaLite notes", (int) length);
     for (int i = 0; i < length; i++) {
         auto event = beatmapEvents->get(i);
